Formularios reactivos, falta editar fruta y el cambio del formulario

+++ angular/src/app/components/pagina-service-frutas/pagina-service-frutas.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Fruta } from 'src/app/model/fruta';
 import { FrutaService } from 'src/app/providers/fruta.service';
-import { Persona } from 'src/app/model/persona';
+import { Mensaje } from 'src/app/model/mensaje';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 
 @Component({
   selector: 'app-pagina-service-frutas',
@@ -9,10 +10,45 @@ import { Persona } from 'src/app/model/persona';
   styleUrls: ['./pagina-service-frutas.component.scss']
 })
 export class PaginaServiceFrutasComponent implements OnInit {
-  frutas:Fruta[];
-  constructor(private frutaService:FrutaService) { 
+  frutas: Fruta[];
+  nuevaFruta: string;
+  frutaSeleccionada: Fruta;
+  mensajeMostrar: Mensaje;
+  formulario: FormGroup; //agrupacion de FormControls ==Input
+
+  constructor(private frutaService: FrutaService, private formBuilder: FormBuilder) {
     console.trace('PaginaServiceFrutasComponent constructor');
-    this.frutas=[];
+    this.frutas = [];
+    this.nuevaFruta = "";
+    this.mensajeMostrar = new Mensaje('', '');
+
+    // inicializamos el formulario
+    this.formulario = this.formBuilder.group({
+      //Aqui se definen los formControll
+      //FormControll --> nombre
+      nombre: [
+        'Fruta Nueva',//value
+        [Validators.required, Validators.minLength(3), Validators.maxLength(150)]//validaciones
+      ],
+      //end control --> nombre
+      precio: [
+        0.99,
+        [Validators.required, Validators.min(0), Validators.max(100)]
+      ],
+      id: [
+        0,
+        [Validators.required, Validators.min(1),]
+      ],
+      oferta: [
+        false,
+        [ Validators.required]
+      ],
+      descuento: [
+        0,
+        [Validators.required, Validators.min(0),Validators.max(100)]
+      ]
+    });
+
   }
 
   ngOnInit() {
@@ -20,26 +56,72 @@ export class PaginaServiceFrutasComponent implements OnInit {
     this.getAllFrutas();
 
   }
-  getAllFrutas(){
-    this.frutaService.getAll().subscribe(jsonData=>{
-      
-      console.debug('Se ha recibido el listado de frutas %o',jsonData);
-      this.frutas = jsonData.map(v=>{
+  getAllFrutas() {
+    this.frutaService.getAll().subscribe(jsonData => {
+
+      console.debug('Se ha recibido el listado de frutas %o', jsonData);
+      this.frutas = jsonData.map(v => {
         return this.rowMapper(v);
       });
     },
-    error=>{
-      console.log("Se ha cometido un error al cargar todas las frutas %o",error);
-    });
+      error => {
+        console.log("Se ha cometido un error al cargar todas las frutas %o", error);
+        this.mensajeMostrar = new Mensaje('Lo siento pero no se ha podido realizar la consulta', Mensaje.DANGER);
+      });
   }
-  
-  delete(fruta:Fruta){
-    this.frutaService.delete(fruta.id).subscribe(result=>{
+
+  delete(fruta: Fruta) {
+    console.trace('PaginaServiceFrutasComponent delete');
+    if (confirm('¿Estas seguro que quieres eliinar?')) {
+
+      this.frutaService.delete(fruta.id).subscribe(result => {
+        this.getAllFrutas();
+        this.mensajeMostrar = new Mensaje(`ELIMINADA ${fruta.nombre}`, 'alert-success');
+      },
+        error => {
+          console.log(error);
+          this.mensajeMostrar = new Mensaje(`No se ha podido ELIMINAR ${fruta.nombre}`, 'alert-warning');
+        });
+    }
+
+  }
+
+  /**
+   * Metodo que inserta una nueva fruta
+   * solamente añade el nombre que le ha dado el usuario,
+   * los demas datos, POR AHORA, están hardcodeados
+   */
+  insert() {
+    console.trace('Crear nueva fruta insert %o', this.formulario.value);
+
+    let frutaNueva = this.rowMapper(this.formulario.value);
+
+    this.frutaService.post(frutaNueva).subscribe(r => {
       this.getAllFrutas();
+      this.mensajeMostrar = new Mensaje(`CREADO ${frutaNueva.nombre}`, 'alert-success');
+      this.nuevaFruta = "";
     },
-    error=>{
-      console.log("Se ha cometido un error al eliminar la fruta seleccionada %o",error);
-    });
+      error => {
+        console.log("Se ha cometido un error al crear la fruta nueva %o", error);
+        this.mensajeMostrar = new Mensaje(`No se ha podido CREAR ${frutaNueva.nombre}`, 'alert-warning');
+        this.nuevaFruta = "";
+      });
+  }
+  update(){
+
+  }
+
+  /**
+   * Metodo que introduce los datos de la fruta seleccionada al formulario para editar dicha fruta
+   * @param fruta  fruta que se ha seleccionado para modificar
+   */
+  seleccionaFrutaEditar(fruta:Fruta){
+    console.trace('click editar %o', fruta);
+    this.formulario.controls['nombre'].setValue(fruta.nombre);
+    this.formulario.controls['precio'].setValue(fruta.precio);
+    this.formulario.controls['id'].setValue(fruta.id);
+    this.formulario.controls['oferta'].setValue(fruta.oferta);
+    this.formulario.controls['descuento'].setValue(fruta.descuento);
   }
 
   /** 
@@ -47,22 +129,26 @@ export class PaginaServiceFrutasComponent implements OnInit {
    * @param value --> es el parametro donde estan los datos de la fruta a transformar
    * @returns Fruta completa
    */
-  rowMapper(value:any):Fruta{
-   return new Fruta(
-    value.nombre,
-    value.precio,
-    value.id,
-    value.oferta,
-    value.descuento,
-    value.imagen,
-      )
+  rowMapper(value: any): Fruta {
+    return new Fruta(
+      value.nombre,
+      value.precio,
+      value.id,
+      value.oferta,
+      value.descuento,
+      value.imagen,
+    )
+  }
+
+  seleccionarFruta(f: Fruta) {
+    this.frutaSeleccionada = f;
   }
 
   /**
    * Metodo que hace el calculo de su precio final con el porcentaje de descuento
    * @param fruta 
    */
-  calcularDescuento(fruta:any):number{
-    return fruta.precio-((fruta.precio*fruta.descuento)/100);
+  calcularDescuento(fruta: any): number {
+    return fruta.precio - ((fruta.precio * fruta.descuento) / 100);
   }
 }
